% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_capture_client.R
\docType{class}
\name{DataCaptureClient-class}
\alias{DataCaptureClient-class}
\alias{DataCaptureClient}
\title{Data Capture Client Class}
\description{
Data Capture Client
}
\details{
Data Capture Client to help capture prediction data
}
\section{Fields}{

\describe{
\item{\code{feature_names}}{array of feature names}

\item{\code{predict_names}}{array of prediction names}

\item{\code{metadata_names}}{array of metadata names}

\item{\code{instance_id}}{string containing a unique instance id}
}}


\examples{
library(DominoDataCapture)
data_capture_client <- DataCaptureClient(
feature_names = c("sepal.length", "sepal.width", "petal.length"),
predict_names = c("variety")
)
data_capture_client$capturePrediction(
feature_values = c(sepal_length, sepal_width, petal_length),
predict_values = c(predicted_value),
event_id = event_id
)
}
